Move Command classes to models package
