+++ data/tundra.txt
@@ -1,0 +1,3 @@
+T | 0 | borrow book
+D | Oct 12 2024 06:00 PM | 1 | return book
+E | Jan 31 2025 08:15 PM | Jan 31 2025 08:30 PM | 0 | project meeting
