+++ src/main/java/tundra/models/ListCommand.java
@@ -1,8 +1,8 @@
-package tundra.views;
+package tundra.models;
 
-import tundra.models.Task;
 import tundra.utils.Storage;
 import tundra.utils.TaskList;
+import tundra.views.Ui;
 
 import java.util.ArrayList;
