+++ src/main/java/tundra/models/TodoCommand.java
@@ -1,10 +1,9 @@
-package tundra.views;
+package tundra.models;
 
 import tundra.exceptions.TundraException;
-import tundra.models.Task;
-import tundra.models.TodoTask;
 import tundra.utils.Storage;
 import tundra.utils.TaskList;
+import tundra.views.Ui;
 
 public class TodoCommand extends Command {
