+++ src/main/java/tundra/models/UnmarkCommand.java
@@ -1,9 +1,9 @@
-package tundra.views;
+package tundra.models;
 
 import tundra.exceptions.TundraException;
-import tundra.models.Task;
 import tundra.utils.Storage;
 import tundra.utils.TaskList;
+import tundra.views.Ui;
 
 public class UnmarkCommand extends Command {
